requirements.txt
flask
requests
flask-mongoengine
flask-login
flask-principal
authlib
loginpass
passlib
bcrypt
pygments
markdown
python-markdown-math
email_validator
wheel
qrcode
pillow
pypng
pandas
xlsxwriter
dateutils

brython
watchdog
